Fix bugs and organize tests
