+++ pkg/tests/command_test.go
@@ -1,4 +1,4 @@
-package tests
+package pkg
 
 import (
 	"bytes"
@@ -13,32 +13,14 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func MockCommands() []pkg.Command {
-	cmds := []pkg.Command{
-		{
-			CmdType:  pkg.ReactiveActuator,
-			Id:       "abc",
-			Cmd:      1,
-			GardenId: "bcd",
-		},
-		{
-			CmdType:  pkg.ScheduledActuator,
-			Id:       "aab",
-			Cmd:      0,
-			GardenId: "bcd",
-		},
-	}
-	return cmds
-}
-
 func TestCreateRACommands(t *testing.T) {
-	pkg.Cache = mockRaCache()
-	cmds, err := pkg.CreateRACommands(mockGardenConfig())
+	pkg.Cache = mocks.MockRaCache()
+	cmds, err := pkg.CreateRACommands(mocks.MockGardenConfig())
 	expectedCmd := pkg.Command{
 		CmdType:  pkg.ReactiveActuator,
-		Id:       raId,
+		Id:       mocks.RAId,
 		Cmd:      1,
-		GardenId: gardenId,
+		GardenId: mocks.GardenId,
 	}
 	assert.NoError(t, err)
 	assert.Equal(t, len(cmds), 1)
@@ -50,9 +32,9 @@ func TestUpdateRACache(t *testing.T) {
 	defer ctrl.Finish()
 	mockStorage := mocks.NewMockStorage(ctrl)
 	mockClient := mocks.NewMockHTTPClient(ctrl)
-	raConfigs := mockGardenConfig().RAConfigs
+	raConfigs := mocks.MockGardenConfig().RAConfigs
 
-	b, err := json.Marshal(mockRA())
+	b, err := json.Marshal(mocks.MockRA())
 	assert.NoError(t, err)
 	r := io.NopCloser(bytes.NewReader(b))
 	res := http.Response{
@@ -61,17 +43,17 @@ func TestUpdateRACache(t *testing.T) {
 	}
 	mockStorage.EXPECT().CreateRAReq(raConfigs[0].RAId).Return(&http.Request{}, nil)
 	mockClient.EXPECT().Do(&http.Request{}).Return(&res, nil)
-	b2, err := json.Marshal(mockSensorLogs())
+	b2, err := json.Marshal(mocks.MockSensorLogs())
 	assert.NoError(t, err)
 	r2 := io.NopCloser(bytes.NewReader(b2))
 	res2 := http.Response{
 		StatusCode: 200,
 		Body:       r2,
 	}
-	mockStorage.EXPECT().CreateSensorLogsReq(sensorId, "1").Return(&http.Request{}, nil)
+	mockStorage.EXPECT().CreateSensorLogsReq(mocks.SensorId, "1").Return(&http.Request{}, nil)
 	mockClient.EXPECT().Do(&http.Request{}).Return(&res2, nil)
-	mockStorage.EXPECT().CreateRALogsReq(raId, "1").Return(&http.Request{}, nil)
-	b3, err := json.Marshal(mockRaLogs())
+	mockStorage.EXPECT().CreateRALogsReq(mocks.RAId, "1").Return(&http.Request{}, nil)
+	b3, err := json.Marshal(mocks.MockRaLogs())
 	assert.NoError(t, err)
 	r3 := io.NopCloser(bytes.NewReader(b3))
 	res3 := http.Response{
@@ -79,7 +61,7 @@ func TestUpdateRACache(t *testing.T) {
 		Body:       r3,
 	}
 	mockClient.EXPECT().Do(&http.Request{}).Return(&res3, nil)
-	cache, err := pkg.UpdateRACache(pkg.Cache, raConfigs, mockRaCache().GardenId, mockClient, mockStorage)
+	cache, err := pkg.UpdateRACache(pkg.Cache, raConfigs, mocks.MockRaCache().GardenId, mockClient, mockStorage)
 	assert.NoError(t, err)
-	assert.Equal(t, mockRaCache(), cache)
+	assert.Equal(t, mocks.MockRaCache(), cache)
 }
